+++ src/components/GMap/Pin.js
@@ -1,23 +1,11 @@
 import React, { Component, PropTypes } from 'react';
 import shouldPureComponentUpdate from 'react-pure-render/function';
-import { connect } from 'react-redux';
-import { actions as coworksActions } from 'redux/modules/coworks/coworks';
 import style from './Pin.scss';
-
-
-const mapStateToProps = (state) => {
-  return {
-    selectedCowork: state.coworks.selected,
-    hoveredCowork: state.coworks.hovered
-  };
-};
 
 class MyPin extends Component {
   static propTypes = {
     text: PropTypes.any.isRequired,
-    identificator: PropTypes.any.isRequired,
-    selectedCowork: PropTypes.number.isRequired,
-    hoveredCowork: PropTypes.number.isRequired
+    identificator: PropTypes.any.isRequired
   };
 
   constructor (props) {
@@ -39,4 +27,4 @@ class MyPin extends Component {
   }
 }
 
-export default connect(mapStateToProps, coworksActions)(MyPin);
+export default MyPin;