+++ src/components/Loading/Loading.js
@@ -1,21 +1,16 @@
-import React, { Component, PropTypes } from 'react';
-import style from './Loading.scss';
-
-const mapStateToProps = (state) => {
-  return {
-    isShown: state.sideBar.isShown,
-    isDocked: state.sideBar.isDocked
-  };
-};
+import React, { Component } from 'react';
 
 class Loading extends Component {
   render () {
     return (
-      <center>
-        <h1><i className='fa fa-spinner fa-spin' /> </h1>
-        <h1> Loading... </h1>
-      </center>
+      <div>
+        <center>
+          <h1><i className='fa fa-spinner fa-spin' /> </h1>
+          <h1> Loading... </h1>
+        </center>
+      </div>
     );
   }
 }
+export { Loading };
 export default Loading;