+++ src/components/UserProfile/UserProfile.js
@@ -1,5 +1,4 @@
 import React, { Component, PropTypes } from 'react';
-import Sidebar from 'react-sidebar';
 import { connect } from 'react-redux';
 import { actions as sideBarActions } from '../../redux/modules/sideBar';
 
@@ -14,7 +13,7 @@ class SideBar extends Component {
   static propTypes = {
     image: PropTypes.string,
     displayName: PropTypes.string
-  }
+  };
 
   componentDidMount () {
   }