+++ src/containers/Root.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import { Provider } from 'react-redux';
 import { Router } from 'react-router';
-import { Notifications, NavBar } from 'components';
+import { Notifications } from 'components';
 
 export default class Root extends React.Component {
   static propTypes = {
     history: React.PropTypes.object.isRequired,
     routes: React.PropTypes.element.isRequired,
     store: React.PropTypes.object.isRequired
-  }
+  };
 
   render () {
     const content = (
@@ -34,7 +34,9 @@ export default class Root extends React.Component {
     }
     return (
       <Provider store={this.props.store}>
-        {content}
+        <div>
+          {content}
+        </div>
       </Provider>
     );
   }